feat: encapsulate profile logic in GeneralInformation
